A couple minor changes to get tests to run and fail usefully.
